@@ -1,29 +1,35 @@
 {
+  "homepage": "http://Anuragkumarsah.github.io/Crypto-World",
   "name": "crypto-tracker",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "axios": "^0.21.1",
-    "chart.js": "^3.5.1",
-    "react": "^17.0.2",
-    "react-alice-carousel": "^2.5.1",
-    "react-chartjs-2": "^3.0.4",
-    "react-dom": "^17.0.2",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/lab": "^4.0.0-alpha.61",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "assert": "^2.0.0",
+    "axios": "^1.1.3",
+    "buffer": "^6.0.3",
+    "chart.js": "^3.9.1",
+    "gh-pages": "^4.0.0",
+    "react": "^18.2.0",
+    "react-alice-carousel": "^2.6.4",
+    "react-chartjs-2": "^4.3.1",
+    "react-dom": "^18.2.0",
     "react-html-parser": "^2.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.0.1"
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^3.0.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build",
+    "build": "react-scripts build && cp build/index.html build/404.html"
   },
   "eslintConfig": {
     "extends": [
